requirements.txt
# Python dependencies for local development and testing
apache-airflow==2.7.0
apache-airflow-providers-google>=10.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.13.0
pandas>=1.3.0
python-dateutil>=2.8.0
pydantic>=1.10.0
